Add map system with grid, seed tray, and collectables
Consolidate map interfaces and classes from interfaces/map/ + classes/map/
into Map/. Includes GridManager/GridPlot/Lane for tile-based grid,
SeedSlot tray with seed packet rendering, Map class orchestrating
background/UI/plant placement, Shovel, Lawnmower, SeedPacket, and
Sun/Coin collectables.

Made-with: Cursor
