+++ Map/CoinCollectable.cs
@@ -1,4 +1,3 @@
-using System.Drawing;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 
@@ -6,16 +5,16 @@ public class CoinCollectable : ICollectable
 {
     public int Value { get; private set; }
     public bool IsCollected { get; private set; }
-    public System.Drawing.Point Position { get; set; }
+    public Point Position { get; set; }
     public float LifetimeRemaining { get; private set; }
-    public System.Drawing.Rectangle Bounds { get; private set; }
+    public Rectangle Bounds { get; private set; }
     public int DrawOrder { get; set; }
 
     private float lifetime;
     private float fallSpeed;
     private bool isFalling;
 
-    public CoinCollectable(System.Drawing.Point position, int value, float lifetime, System.Drawing.Rectangle bounds)
+    public CoinCollectable(Point position, int value, float lifetime, Rectangle bounds)
     {
         Position = position;
         Value = value;
@@ -28,12 +27,12 @@ public class CoinCollectable : ICollectable
         fallSpeed = 50f;
     }
 
-    public bool HitTest(System.Drawing.Point mousePos)
+    public bool HitTest(Point mousePos)
     {
         return Bounds.Contains(mousePos);
     }
 
-    public void OnClick(MouseController mouse)
+    public void OnClick(IMouse mouse)
     {
         Collect();
     }
@@ -42,7 +41,6 @@ public class CoinCollectable : ICollectable
     {
         if (!IsCollected)
         {
-            // TODO: Draw coin sprite at Position
         }
     }
 
@@ -54,8 +52,8 @@ public class CoinCollectable : ICollectable
 
         if (isFalling)
         {
-            Position = new System.Drawing.Point(Position.X, Position.Y + (int)(fallSpeed * deltaTime));
-            Bounds = new System.Drawing.Rectangle(
+            Position = new Point(Position.X, Position.Y + (int)(fallSpeed * deltaTime));
+            Bounds = new Rectangle(
                 Position.X - Bounds.Width / 2,
                 Position.Y - Bounds.Height / 2,
                 Bounds.Width, Bounds.Height);
