+++ Map/IShovel.cs
@@ -1,0 +1,8 @@
+public interface IShovel : IClickable, IPvZDrawable
+{
+    bool IsSelected { get; set; }
+
+    void Activate();
+    void Deactivate();
+    void UseOnPlot(IGridPlot plot);
+}
